requirements.txt
agno==1.7.7
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
docstring-parser==0.17.0
executing==2.2.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.40.3
google-genai==1.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mcp==1.12.3
mdurl==0.1.2
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
pyyaml==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
tenacity==8.5.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
